url-ailoader/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <mr.junaidshaukat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.8"
langchain = "^0.1.11"
playwright = "^1.42.0"
beautifulsoup4 = "^4.12.3"
asyncio = "^3.4.3"
sqlmodel = "^0.0.16"
pydantic = "^2.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
